requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
PyYAML==6.0.2
httpx==0.27.0
markdown==3.6
